The provided commit message does not match with the given code differences. Here's the proper commit message for these code changes:
Commit message:
Add script to clone tools repo and update .gitignore

The `init.sh` script has been modified to include a check for the existence of a 'tools' directory and to clone `examples_tools` repo if it does not exist. The `.gitignore` file has also been updated to ignore the 'tools' directory.